refactor: moved eyca saga into separate file
