+++ apps/main-app/ts/features/bonus/cgn/saga/__tests__/orchestration/eyca/activation/eycaActivationSaga.test.ts
@@ -4,12 +4,10 @@ import { testSaga } from "redux-saga-test-plan";
 import { getGenericError } from "../../../../../../../../utils/errors";
 import { cgnEycaActivation } from "../../../../../store/actions/eyca/activation";
 import { cgnEycaStatus } from "../../../../../store/actions/eyca/details";
-import {
-  getActivation,
-  handleEycaActivationSaga,
-  handleStartActivation
-} from "../../../../networking/eyca/activation/getEycaActivationSaga";
-import { eycaActivationWorker } from "../../../../orchestration/eyca/eycaActivationSaga";
+import { getActivation } from "../../../../networking/eyca/activation/getActivation";
+import { handleEycaActivationSaga } from "../../../../networking/eyca/activation/handleEycaActivationSaga";
+import { handleStartActivation } from "../../../../networking/eyca/activation/handleStartActivation";
+import { eycaActivationWorker } from "../../../../orchestration/eyca/eycaActivationWorker";
 import {
   navigateToCgnDetails,
   navigateToEycaActivationLoading
