+++ apps/main-app/ts/features/bonus/cgn/saga/networking/eyca/activation/__tests__/getEycaActivationStatus.test.ts
@@ -7,7 +7,7 @@ import {
 } from "../../../../../../../../utils/errors";
 import { cgnEycaActivation } from "../../../../../store/actions/eyca/activation";
 import { CgnEycaActivationStatus } from "../../../../../store/reducers/eyca/activation";
-import { getActivation } from "../getEycaActivationSaga";
+import { getActivation } from "../getActivation";
 import { getEycaActivationStatusSaga } from "../getEycaActivationStatus";
 
 describe("getEycaActivationStatus", () => {
