+++ apps/main-app/ts/features/bonus/cgn/saga/orchestration/eyca/eycaActivationWorker.ts
@@ -1,0 +1,70 @@
+import * as E from "fp-ts/lib/Either";
+import { call, put } from "typed-redux-saga/macro";
+
+import { SagaCallReturnType } from "../../../../../../types/utils";
+import { BackendCGN } from "../../../api/backendCgn";
+import { cgnEycaActivation } from "../../../store/actions/eyca/activation";
+import { cgnEycaStatus } from "../../../store/actions/eyca/details";
+import { getActivation } from "../../networking/eyca/activation/getActivation";
+import { handleEycaActivationSaga } from "../../networking/eyca/activation/handleEycaActivationSaga";
+import { handleStartActivation } from "../../networking/eyca/activation/handleStartActivation";
+import {
+  navigateToCgnDetails,
+  navigateToEycaActivationLoading
+} from "../navigation/actions";
+
+/**
+ * This saga handles the activation request for an EYCA Card linked to the user's CGN.
+ * This activation flow may be invoked only if the background activation process originated from the Backend
+ * terminated with errors and no EYCA information is available.
+ * First we check the current status of the activation orchestrator and then we follow this logic:
+ * - COMPLETED/ERROR/NOT_FOUND: request for a new activation, then start polling until we have results
+ * - PROCESSING/RUNNING: we start polling waiting for results
+ *
+ * At the end navigate back to CGN detail and request again for EYCA Details
+ * @param getEycaActivation
+ * @param startEycaActivation
+ */
+export function* eycaActivationWorker(
+  getEycaActivation: ReturnType<typeof BackendCGN>["getEycaActivation"],
+  startEycaActivation: ReturnType<typeof BackendCGN>["startEycaActivation"]
+) {
+  yield* call(navigateToEycaActivationLoading);
+
+  const eycaActivation: SagaCallReturnType<typeof getActivation> = yield* call(
+    getActivation,
+    getEycaActivation
+  );
+
+  if (E.isRight(eycaActivation)) {
+    if (eycaActivation.right === "PROCESSING") {
+      yield* call(handleEycaActivationSaga, getEycaActivation);
+    } else {
+      const startActivation: SagaCallReturnType<typeof handleStartActivation> =
+        yield* call(handleStartActivation, startEycaActivation);
+      // activation not handled error, stop
+      if (E.isLeft(startActivation)) {
+        yield* put(cgnEycaActivation.failure(startActivation.left));
+        return;
+      } else {
+        // could be: ALREADY_ACTIVE, INELIGIBLE
+        if (
+          ["ALREADY_ACTIVE", "INELIGIBLE"].some(
+            v => v === startActivation.right
+          )
+        ) {
+          yield* put(cgnEycaActivation.success(startActivation.right));
+          yield* call(navigateToCgnDetails);
+          return;
+        } else {
+          yield* call(handleEycaActivationSaga, getEycaActivation);
+        }
+      }
+    }
+  }
+
+  // Activation saga ended, request again the details
+  yield* put(cgnEycaStatus.request());
+
+  yield* call(navigateToCgnDetails);
+}
